Move files and use feature flag on submit loader
