+++ frontend/src/app/testQueue/QueueItemSubmitLoader.scss
@@ -31,3 +31,10 @@
     display: none;
   }
 }
+
+.test-card-submit-loader {
+  top: 0;
+  left: 0;
+  border-bottom-left-radius: 0.5rem !important;
+  border-bottom-right-radius: 0.5rem !important;
+}
