+++ frontend/src/app/testQueue/QueueItemSubmitLoader.tsx
@@ -4,6 +4,7 @@ import { CSSTransition } from "react-transition-group";
 import iconLoader from "../../img/loader.svg";
 
 import "./QueueItemSubmitLoader.scss";
+import { useFeature } from "flagged";
 
 type Props = {
   name: string;
@@ -11,6 +12,10 @@ type Props = {
 };
 
 export const QueueItemSubmitLoader = ({ name, show }: Props) => {
+  const testCardRefactorEnabled = useFeature(
+    "testCardRefactorEnabled"
+  ) as boolean;
+
   const classnames = classNames(
     "sr-queue-item-submit-loader",
     "z-top",
@@ -18,8 +23,12 @@ export const QueueItemSubmitLoader = ({ name, show }: Props) => {
     "height-full",
     "width-full",
     "text-center",
-    "radius-lg"
+    testCardRefactorEnabled ? "test-card-submit-loader" : "radius-lg"
   );
+
+  const headingClassNames = testCardRefactorEnabled
+    ? ""
+    : classNames("margin-top-6", "margin-bottom-5");
 
   return (
     <CSSTransition
@@ -28,7 +37,7 @@ export const QueueItemSubmitLoader = ({ name, show }: Props) => {
       classNames="sr-queue-item-submit-loader"
     >
       <div className={classnames} aria-hidden={!show}>
-        <h4 className="margin-top-6 margin-bottom-5">
+        <h4 className={headingClassNames}>
           Submitting test data for {name}...
         </h4>
         <img src={iconLoader} alt="submitting" className="square-5" />
