+++ frontend/src/app/testQueue/TestCardForm/diseaseSpecificComponents/CovidResultInputGroup.tsx
